Move test helper to test folder.
