+++ streaming/java/streaming-runtime/src/test/java/org/ray/streaming/runtime/TestHelper.java
@@ -1,4 +1,4 @@
-package org.ray.streaming.runtime.util;
+package org.ray.streaming.runtime;
 
 public class TestHelper {
